refactor(bridge): move createServerSidePublicClient into the bridge package
It lived under earn (@/earn-api), forcing the CCTP backfill (a bridge module) to import from the app package, a backwards dependency. Move it to util/rpc/serverPublicClient.ts and point both earn and the CCTP backfill at it.
